+++ expo-example/App.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import Box from 'rebass-native';
+export default class App extends React.Component {
+  render() {
+    return (
+      <Box width={100} height={100} bg="hotpink"/>
+    );
+  }
+}