+++ src/BackgroundImage.js
@@ -1,4 +1,4 @@
-import sys from 'system-components'
+import sys from 'native-system-components'
 import { style } from 'styled-system'
 
 const bgImage = style({