+++ src/Close.js
@@ -1,4 +1,4 @@
-import sys from 'system-components'
+import sys from 'native-system-components'
 import ButtonTransparent from './ButtonTransparent'
 
 export const Close = sys({