+++ src/Toolbar.js
@@ -1,5 +1,5 @@
-import sys from 'system-components'
-import { Flex } from 'grid-styled'
+import sys from 'native-system-components'
+import { Flex } from 'native-grid-styled'
 
 export const Toolbar = sys({
   is: Flex,